feat: #4107 account detail add resource tenant info
