+++ containers/Cloudenv/views/cloudprovider/mixins/columns.js
@@ -5,6 +5,7 @@ import {
   getCopyWithContentTableColumn,
 } from '@/utils/common/tableColumn'
 import i18n from '@/locales'
+import { getResourceMatchProjectTableColumn } from '../utils/columns'
 
 export default {
   created () {
@@ -68,28 +69,7 @@ export default {
         field: 'project_domain',
         title: i18n.t('cloudenv.text_355', [i18n.t('dictionary.domain')]),
       }),
-      {
-        field: 'tenant',
-        title: i18n.t('cloudenv.text_356', [i18n.t('dictionary.project')]),
-        minWidth: 120,
-        showOverflow: 'title',
-        slots: {
-          default: ({ row }) => {
-            const ret = []
-            ret.push(<list-body-cell-wrap copy field='tenant' row={row} />)
-            if (row.project_mapping) {
-              let label = ''
-              if (row.enable_resource_sync) {
-                label = i18n.t('cloudenv.resource_project_mapping')
-              } else if (row.enable_project_sync) {
-                label = i18n.t('cloudenv.project_project_mapping')
-              }
-              ret.push(<list-body-cell-wrap copy field='project_mapping' row={row} hideField><span class="text-color-secondary">{label || this.$t('cloudenv.text_580')}：{row.project_mapping}</span></list-body-cell-wrap>)
-            }
-            return ret
-          },
-        },
-      },
+      getResourceMatchProjectTableColumn(),
     ]
   },
 }
