+++ containers/Cloudenv/views/cloudprovider/sidepage/Detail.vue
@@ -13,6 +13,9 @@ import {
   getEnabledTableColumn,
   getBrandTableColumn,
 } from '@/utils/common/tableColumn'
+import {
+  getResourceMatchProjectTableColumn,
+} from '../utils/columns'
 export default {
   name: 'CloudproviderDetail',
   props: {
@@ -28,6 +31,7 @@ export default {
   data () {
     return {
       baseInfo: [
+        getResourceMatchProjectTableColumn(),
         getBrandTableColumn({ field: 'provider' }),
         {
           field: 'account',
